nvd_data_extract/requirements.txt
pymysql
sqlalchemy>=1.4.*
requests>=2.31.*
google-cloud-storage>=2.14.*
cloud-sql-python-connector>=1.6.*
google-api-python-client>=1.7.12
functions-framework==3.*
google-cloud-secret-manager